RavenDB-22429 Rename tests
